+++ src/formBuilder/utils.js
@@ -156,6 +156,7 @@ const supportedUiParameters = new Set([
   'ui:autocomplete',
   'ui:options',
   'ui:field',
+  'ui:placeholder',
   'ui:column',
   'items',
   'definitions',
@@ -355,7 +356,7 @@ export function checkForUnsupportedFeatures(
 }
 
 // make an element out of the corresponding properties and UI properties
-function generateElement(
+function generateDependencyElement(
   name: string,
   dataProps: any,
   uiProperties: any,
@@ -363,6 +364,7 @@ function generateElement(
   definitionData?: { [string]: any },
   definitionUi?: { [string]: any },
   categoryHash: { [string]: string },
+  useDefinitionDetails: boolean = true, // determines whether to use an element's definition details or not.
 ) {
   let uiProps = {
     ...uiProperties,
@@ -380,7 +382,8 @@ function generateElement(
     if (
       pathArr[0] === '#' &&
       pathArr[1] === 'definitions' &&
-      definitionData[pathArr[2]]
+      definitionData[pathArr[2]] &&
+      useDefinitionDetails === true
     ) {
       elementDetails = {
         ...elementDetails,
@@ -479,7 +482,6 @@ export function generateElementPropsFromSchemas(parameters: {
         ...uischema[parameter],
       };
     }
-
     newElement.name = parameter;
     newElement.required = requiredNames.includes(parameter);
     newElement.$ref = elementDetails.$ref;
@@ -512,6 +514,7 @@ export function generateElementPropsFromSchemas(parameters: {
   });
   // read dependent elements from dependencies
   if (schema.dependencies) {
+    const useDefinitionDetails = false;
     Object.keys(schema.dependencies).forEach((parent) => {
       const group = schema.dependencies[parent];
       if (group.oneOf) {
@@ -529,7 +532,7 @@ export function generateElementPropsFromSchemas(parameters: {
             ([parameter, element]) => {
               // create a new element if not present in main properties
               if (!Object.keys(elementDict).includes(parameter)) {
-                const newElement = generateElement(
+                const newElement = generateDependencyElement(
                   parameter,
                   element,
                   uischema[parameter],
@@ -537,6 +540,7 @@ export function generateElementPropsFromSchemas(parameters: {
                   definitionData,
                   definitionUi,
                   categoryHash,
+                  useDefinitionDetails,
                 );
                 newElement.required = requiredValues.includes(newElement.name);
                 elementDict[newElement.name] = newElement;
@@ -556,7 +560,7 @@ export function generateElementPropsFromSchemas(parameters: {
       } else if (group.properties) {
         const requiredValues = group.required || [];
         Object.entries(group.properties).forEach(([parameter, element]) => {
-          const newElement = generateElement(
+          const newElement = generateDependencyElement(
             parameter,
             element,
             uischema[parameter],
@@ -564,6 +568,7 @@ export function generateElementPropsFromSchemas(parameters: {
             definitionData,
             definitionUi,
             categoryHash,
+            useDefinitionDetails,
           );
           newElement.required = requiredValues.includes(newElement.name);
           newElement.dependent = true;
@@ -808,7 +813,8 @@ export function generateUiSchemaFromElementProps(
         typeof element.$ref === 'string' ? element.$ref.split('/') : [];
       if (definitions && definitions[pathArr[2]])
         uiSchema[element.name] = definitions[pathArr[2]];
-    } else if (element.propType === 'card' && element.uiOptions) {
+    }
+    if (element.propType === 'card' && element.uiOptions) {
       Object.keys(element.uiOptions).forEach((uiOption) => {
         if (!uiSchema[element.name]) uiSchema[element.name] = {};
         if (uiOption.startsWith('ui:*')) {